102 feature phased pull request merge (#109)
* add music module in Design4Version1.0.0.drawio

* update docs

* add metadata search api but not impl

* impl metadata search api in MetadataServiceImpl.java

* add subscribe entity and repository and service

* optimize season episode and file matching logic

* fix upload issue when file repeated

* optimize date parse fail log print

* fix some issue for load rest template proxy status after app start

* fix update episode url with file id in TaskServiceImpl#createServerFileHardLink

* change not update when episode url exists

* add logic that server file not create hard link when file has exists by name

* separate generate media dir from task service to media server

* fix app init fail issue

* update admin page for add media generate switch in system option tab

* fix generate media dir air date incorrect issue

* update CHANGELOG.MD

+++ src/main/java/run/ikaros/server/service/AbstractCrudService.java
@@ -3,10 +3,12 @@ package run.ikaros.server.service;
 import java.util.List;
 import java.util.Optional;
 import javax.annotation.Nonnull;
+
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.lang.Nullable;
 import run.ikaros.server.core.service.CrudService;
 import run.ikaros.server.entity.BaseEntity;
 import run.ikaros.server.exceptions.RecordNotFoundException;
@@ -59,6 +61,9 @@ public class AbstractCrudService<E, I> implements CrudService<E, I> {
     public E deleteById(@Nonnull I id) {
         AssertUtils.notNull(id, "id");
         E e = getById(id);
+        if (e == null) {
+            return e;
+        }
         baseRepository.delete(e);
         return e;
     }
@@ -67,6 +72,9 @@ public class AbstractCrudService<E, I> implements CrudService<E, I> {
     @SuppressWarnings("unchecked")
     public E removeById(@Nonnull I id) {
         E e = getById(id);
+        if (e == null) {
+            return e;
+        }
         BaseEntity entity = (BaseEntity) e;
         entity.setStatus(false);
         e = (E) entity;
@@ -115,12 +123,11 @@ public class AbstractCrudService<E, I> implements CrudService<E, I> {
         return baseRepository.findById(id);
     }
 
-    @Nonnull
+    @Nullable
     @Override
     public E getById(@Nonnull I id) {
         AssertUtils.notNull(id, "id");
-        return baseRepository.findById(id).orElseThrow(
-            () -> new RecordNotFoundException("record not found, id=" + id));
+        return baseRepository.findById(id).orElse(null);
     }
 
     @Override
